chanstate: move kv channel transition test
Move ChannelStateTransition coverage from channeldb into chanstate because the test exercises commitment, revocation-log, and forwarding-package persistence owned by the KV channel-state store.

The moved test uses KVStore directly and replaces the ChannelStateDB revocation-log helper with the chanstate store method.
