+++ channeldb/channel_test.go
@@ -5,8 +5,6 @@ import (
 	"encoding/hex"
 	"math/rand"
 	"net"
-	"reflect"
-	"runtime"
 	"sync/atomic"
 	"testing"
 
@@ -16,10 +14,8 @@ import (
 	"github.com/btcsuite/btcd/chainhash/v2"
 	"github.com/btcsuite/btcd/wire/v2"
 	_ "github.com/btcsuite/btcwallet/walletdb/bdb"
-	"github.com/davecgh/go-spew/spew"
 	"github.com/lightningnetwork/lnd/clock"
 	"github.com/lightningnetwork/lnd/fn/v2"
-	"github.com/lightningnetwork/lnd/graph/db/models"
 	"github.com/lightningnetwork/lnd/keychain"
 	"github.com/lightningnetwork/lnd/kvdb"
 	"github.com/lightningnetwork/lnd/lntest/channels"
@@ -402,370 +398,4 @@ func createTestChannelState(t *testing.T, cdb *ChannelStateDB) *OpenChannel {
 		TapscriptRoot:           fn.Some(tapscriptRoot),
 		CustomBlob:              fn.Some([]byte{1, 2, 3}),
 	}
-}
-
-func assertCommitmentEqual(t *testing.T, a, b *ChannelCommitment) {
-	if !reflect.DeepEqual(a, b) {
-		_, _, line, _ := runtime.Caller(1)
-		t.Fatalf("line %v: commitments don't match: %v vs %v",
-			line, spew.Sdump(a), spew.Sdump(b))
-	}
-}
-
-// assertRevocationLogEntryEqual asserts that, for all the fields of a given
-// revocation log entry, their values match those on a given ChannelCommitment.
-func assertRevocationLogEntryEqual(t *testing.T, c *ChannelCommitment,
-	r *RevocationLog) {
-
-	t.Helper()
-
-	// Check the common fields.
-	require.EqualValues(
-		t, r.CommitTxHash.Val, c.CommitTx.TxHash(), "CommitTx mismatch",
-	)
-
-	// Now check the common fields from the HTLCs.
-	require.Equal(t, len(r.HTLCEntries), len(c.Htlcs), "HTLCs len mismatch")
-	for i, rHtlc := range r.HTLCEntries {
-		cHtlc := c.Htlcs[i]
-		require.Equal(t, rHtlc.RHash.Val[:], cHtlc.RHash[:], "RHash")
-		require.Equal(
-			t, rHtlc.Amt.Val.Int(), cHtlc.Amt.ToSatoshis(), "Amt",
-		)
-		require.Equal(
-			t, rHtlc.RefundTimeout.Val, cHtlc.RefundTimeout,
-			"RefundTimeout",
-		)
-		require.EqualValues(
-			t, rHtlc.OutputIndex.Val, cHtlc.OutputIndex,
-			"OutputIndex",
-		)
-		require.Equal(
-			t, rHtlc.Incoming.Val, cHtlc.Incoming, "Incoming",
-		)
-	}
-}
-
-func TestChannelStateTransition(t *testing.T) {
-	t.Parallel()
-
-	fullDB, err := MakeTestDB(t)
-	require.NoError(t, err, "unable to make test database")
-
-	cdb := fullDB.ChannelStateDB()
-
-	// First create a minimal channel, then perform a full sync in order to
-	// persist the data.
-	channel := createTestChannel(t, cdb)
-
-	// Add some HTLCs which were added during this new state transition.
-	// Half of the HTLCs are incoming, while the other half are outgoing.
-	var (
-		htlcs   []HTLC
-		htlcAmt lnwire.MilliSatoshi
-	)
-	for i := uint32(0); i < 10; i++ {
-		var incoming bool
-		if i > 5 {
-			incoming = true
-		}
-		htlc := HTLC{
-			Signature:     testSig.Serialize(),
-			Incoming:      incoming,
-			Amt:           10,
-			RHash:         key,
-			RefundTimeout: i,
-			OutputIndex:   int32(i * 3),
-			LogIndex:      uint64(i * 2),
-			HtlcIndex:     uint64(i),
-		}
-		copy(
-			htlc.OnionBlob[:],
-			bytes.Repeat([]byte{2}, lnwire.OnionPacketSize),
-		)
-		htlcs = append(htlcs, htlc)
-		htlcAmt += htlc.Amt
-	}
-
-	// Create a new channel delta which includes the above HTLCs, some
-	// balance updates, and an increment of the current commitment height.
-	// Additionally, modify the signature and commitment transaction.
-	newSequence := uint32(129498)
-	newSig := bytes.Repeat([]byte{3}, 71)
-	newTx := channel.LocalCommitment.CommitTx.Copy()
-	newTx.TxIn[0].Sequence = newSequence
-	commitment := ChannelCommitment{
-		CommitHeight:    1,
-		LocalLogIndex:   2,
-		LocalHtlcIndex:  1,
-		RemoteLogIndex:  2,
-		RemoteHtlcIndex: 1,
-		LocalBalance:    lnwire.MilliSatoshi(1e8),
-		RemoteBalance:   lnwire.MilliSatoshi(1e8),
-		CommitFee:       55,
-		FeePerKw:        99,
-		CommitTx:        newTx,
-		CommitSig:       newSig,
-		Htlcs:           htlcs,
-		CustomBlob:      fn.Some([]byte{4, 5, 6}),
-	}
-
-	// First update the local node's broadcastable state and also add a
-	// CommitDiff remote node's as well in order to simulate a proper state
-	// transition.
-	unsignedAckedUpdates := []LogUpdate{
-		{
-			LogIndex: 2,
-			UpdateMsg: &lnwire.UpdateAddHTLC{
-				ChanID: lnwire.ChannelID{1, 2, 3},
-			},
-		},
-	}
-
-	_, err = channel.UpdateCommitment(&commitment, unsignedAckedUpdates)
-	require.NoError(t, err, "unable to update commitment")
-
-	// Assert that update is correctly written to the database.
-	dbUnsignedAckedUpdates, err := channel.UnsignedAckedUpdates()
-	require.NoError(t, err, "unable to fetch dangling remote updates")
-	if len(dbUnsignedAckedUpdates) != 1 {
-		t.Fatalf("unexpected number of dangling remote updates")
-	}
-	if !reflect.DeepEqual(
-		dbUnsignedAckedUpdates[0], unsignedAckedUpdates[0],
-	) {
-
-		t.Fatalf("unexpected update: expected %v, got %v",
-			spew.Sdump(unsignedAckedUpdates[0]),
-			spew.Sdump(dbUnsignedAckedUpdates))
-	}
-
-	// The balances, new update, the HTLCs and the changes to the fake
-	// commitment transaction along with the modified signature should all
-	// have been updated.
-	updatedChannel, err := cdb.FetchOpenChannels(channel.IdentityPub)
-	require.NoError(t, err, "unable to fetch updated channel")
-
-	assertCommitmentEqual(
-		t, &commitment, &updatedChannel[0].LocalCommitment,
-	)
-
-	numDiskUpdates, err := updatedChannel[0].CommitmentHeight()
-	require.NoError(t, err, "unable to read commitment height from disk")
-
-	if numDiskUpdates != uint64(commitment.CommitHeight) {
-		t.Fatalf("num disk updates doesn't match: %v vs %v",
-			numDiskUpdates, commitment.CommitHeight)
-	}
-
-	// Attempting to query for a commitment diff should return
-	// ErrNoPendingCommit as we haven't yet created a new state for them.
-	_, err = channel.RemoteCommitChainTip()
-	if err != ErrNoPendingCommit {
-		t.Fatalf("expected ErrNoPendingCommit, instead got %v", err)
-	}
-
-	// To simulate us extending a new state to the remote party, we'll also
-	// create a new commit diff for them.
-	remoteCommit := commitment
-	remoteCommit.LocalBalance = lnwire.MilliSatoshi(2e8)
-	remoteCommit.RemoteBalance = lnwire.MilliSatoshi(3e8)
-	remoteCommit.CommitHeight = 1
-	commitDiff := &CommitDiff{
-		Commitment: remoteCommit,
-		CommitSig: &lnwire.CommitSig{
-			ChanID:    lnwire.ChannelID(key),
-			CommitSig: wireSig,
-			HtlcSigs: []lnwire.Sig{
-				wireSig,
-				wireSig,
-			},
-		},
-		LogUpdates: []LogUpdate{
-			{
-				LogIndex: 1,
-				UpdateMsg: &lnwire.UpdateAddHTLC{
-					ID:     1,
-					Amount: lnwire.NewMSatFromSatoshis(100),
-					Expiry: 25,
-				},
-			},
-			{
-				LogIndex: 2,
-				UpdateMsg: &lnwire.UpdateAddHTLC{
-					ID:     2,
-					Amount: lnwire.NewMSatFromSatoshis(200),
-					Expiry: 50,
-				},
-			},
-		},
-		OpenedCircuitKeys: []models.CircuitKey{},
-		ClosedCircuitKeys: []models.CircuitKey{},
-	}
-	copy(commitDiff.LogUpdates[0].UpdateMsg.(*lnwire.UpdateAddHTLC).PaymentHash[:],
-		bytes.Repeat([]byte{1}, 32))
-	copy(commitDiff.LogUpdates[1].UpdateMsg.(*lnwire.UpdateAddHTLC).PaymentHash[:],
-		bytes.Repeat([]byte{2}, 32))
-	if err := channel.AppendRemoteCommitChain(commitDiff); err != nil {
-		t.Fatalf("unable to add to commit chain: %v", err)
-	}
-
-	// The commitment tip should now match the commitment that we just
-	// inserted.
-	diskCommitDiff, err := channel.RemoteCommitChainTip()
-	require.NoError(t, err, "unable to fetch commit diff")
-	if !reflect.DeepEqual(commitDiff, diskCommitDiff) {
-		t.Fatalf("commit diffs don't match: %v vs %v", spew.Sdump(remoteCommit),
-			spew.Sdump(diskCommitDiff))
-	}
-
-	// We'll save the old remote commitment as this will be added to the
-	// revocation log shortly.
-	oldRemoteCommit := channel.RemoteCommitment
-
-	// Next, write to the log which tracks the necessary revocation state
-	// needed to rectify any fishy behavior by the remote party. Modify the
-	// current uncollapsed revocation state to simulate a state transition
-	// by the remote party.
-	channel.RemoteCurrentRevocation = channel.RemoteNextRevocation
-	newPriv, err := btcec.NewPrivateKey()
-	require.NoError(t, err, "unable to generate key")
-	channel.RemoteNextRevocation = newPriv.PubKey()
-
-	fwdPkg := NewFwdPkg(channel.ShortChanID(), oldRemoteCommit.CommitHeight,
-		diskCommitDiff.LogUpdates, nil)
-
-	err = channel.AdvanceCommitChainTail(
-		fwdPkg, nil, dummyLocalOutputIndex, dummyRemoteOutIndex,
-	)
-	require.NoError(t, err, "unable to append to revocation log")
-
-	// At this point, the remote commit chain should be nil, and the posted
-	// remote commitment should match the one we added as a diff above.
-	if _, err := channel.RemoteCommitChainTip(); err != ErrNoPendingCommit {
-		t.Fatalf("expected ErrNoPendingCommit, instead got %v", err)
-	}
-
-	// We should be able to fetch the channel delta created above by its
-	// update number with all the state properly reconstructed.
-	diskPrevCommit, _, err := channel.FindPreviousState(
-		oldRemoteCommit.CommitHeight,
-	)
-	require.NoError(t, err, "unable to fetch past delta")
-
-	// Check the output indexes are saved as expected.
-	require.EqualValues(
-		t, dummyLocalOutputIndex, diskPrevCommit.OurOutputIndex.Val,
-	)
-	require.EqualValues(
-		t, dummyRemoteOutIndex, diskPrevCommit.TheirOutputIndex.Val,
-	)
-
-	// The two deltas (the original vs the on-disk version) should
-	// identical, and all HTLC data should properly be retained.
-	assertRevocationLogEntryEqual(t, &oldRemoteCommit, diskPrevCommit)
-
-	// The state number recovered from the tail of the revocation log
-	// should be identical to this current state.
-	logTailHeight, err := cdb.revocationLogTailCommitHeight(channel)
-	require.NoError(t, err, "unable to retrieve log")
-	if logTailHeight != oldRemoteCommit.CommitHeight {
-		t.Fatal("update number doesn't match")
-	}
-
-	oldRemoteCommit = channel.RemoteCommitment
-
-	// Next modify the posted diff commitment slightly, then create a new
-	// commitment diff and advance the tail.
-	commitDiff.Commitment.CommitHeight = 2
-	commitDiff.Commitment.LocalBalance -= htlcAmt
-	commitDiff.Commitment.RemoteBalance += htlcAmt
-	commitDiff.LogUpdates = []LogUpdate{}
-	if err := channel.AppendRemoteCommitChain(commitDiff); err != nil {
-		t.Fatalf("unable to add to commit chain: %v", err)
-	}
-
-	fwdPkg = NewFwdPkg(channel.ShortChanID(), oldRemoteCommit.CommitHeight, nil, nil)
-
-	err = channel.AdvanceCommitChainTail(
-		fwdPkg, nil, dummyLocalOutputIndex, dummyRemoteOutIndex,
-	)
-	require.NoError(t, err, "unable to append to revocation log")
-
-	// Once again, fetch the state and ensure it has been properly updated.
-	prevCommit, _, err := channel.FindPreviousState(
-		oldRemoteCommit.CommitHeight,
-	)
-	require.NoError(t, err, "unable to fetch past delta")
-
-	// Check the output indexes are saved as expected.
-	require.EqualValues(
-		t, dummyLocalOutputIndex, diskPrevCommit.OurOutputIndex.Val,
-	)
-	require.EqualValues(
-		t, dummyRemoteOutIndex, diskPrevCommit.TheirOutputIndex.Val,
-	)
-
-	assertRevocationLogEntryEqual(t, &oldRemoteCommit, prevCommit)
-
-	// Once again, state number recovered from the tail of the revocation
-	// log should be identical to this current state.
-	logTailHeight, err = cdb.revocationLogTailCommitHeight(channel)
-	require.NoError(t, err, "unable to retrieve log")
-	if logTailHeight != oldRemoteCommit.CommitHeight {
-		t.Fatal("update number doesn't match")
-	}
-
-	// The revocation state stored on-disk should now also be identical.
-	updatedChannel, err = cdb.FetchOpenChannels(channel.IdentityPub)
-	require.NoError(t, err, "unable to fetch updated channel")
-	if !channel.RemoteCurrentRevocation.IsEqual(updatedChannel[0].RemoteCurrentRevocation) {
-		t.Fatalf("revocation state was not synced")
-	}
-	if !channel.RemoteNextRevocation.IsEqual(updatedChannel[0].RemoteNextRevocation) {
-		t.Fatalf("revocation state was not synced")
-	}
-
-	// At this point, we should have 2 forwarding packages added.
-	fwdPkgs := loadFwdPkgs(
-		t, cdb.backend, NewChannelPackager(channel.ShortChanID()),
-	)
-	require.Len(t, fwdPkgs, 2, "wrong number of forwarding packages")
-
-	// Now attempt to delete the channel from the database.
-	closeSummary := &ChannelCloseSummary{
-		ChanPoint:         channel.FundingOutpoint,
-		RemotePub:         channel.IdentityPub,
-		SettledBalance:    btcutil.Amount(500),
-		TimeLockedBalance: btcutil.Amount(10000),
-		IsPending:         false,
-		CloseType:         RemoteForceClose,
-	}
-	if err := updatedChannel[0].CloseChannel(closeSummary); err != nil {
-		t.Fatalf("unable to delete updated channel: %v", err)
-	}
-
-	// If we attempt to fetch the target channel again, it shouldn't be
-	// found.
-	channels, err := cdb.FetchOpenChannels(channel.IdentityPub)
-	require.NoError(t, err, "unable to fetch updated channels")
-	if len(channels) != 0 {
-		t.Fatalf("%v channels, found, but none should be",
-			len(channels))
-	}
-
-	// Attempting to find previous states on the channel should fail as the
-	// revocation log has been deleted.
-	_, _, err = updatedChannel[0].FindPreviousState(
-		oldRemoteCommit.CommitHeight,
-	)
-	if err == nil {
-		t.Fatal("revocation log search should have failed")
-	}
-
-	// All forwarding packages of this channel has been deleted too.
-	fwdPkgs = loadFwdPkgs(
-		t, cdb.backend, NewChannelPackager(channel.ShortChanID()),
-	)
-	require.Empty(t, fwdPkgs, "no forwarding packages should exist")
 }
